DataStore clean up: remove deprecated methods, rename to KvDataStore
