+++ infrastructure/src/main/java/org/corfudb/infrastructure/datastore/KvDataStore.java
@@ -1,4 +1,4 @@
-package org.corfudb.infrastructure;
+package org.corfudb.infrastructure.datastore;
 
 
 import lombok.AllArgsConstructor;
