+++ test/src/test/java/org/corfudb/infrastructure/DataStoreTest.java
@@ -10,7 +10,8 @@ import java.util.UUID;
 import java.util.function.Consumer;
 
 import org.corfudb.AbstractCorfuTest;
-import org.corfudb.infrastructure.KvDataStore.KvRecord;
+import org.corfudb.infrastructure.datastore.DataStore;
+import org.corfudb.infrastructure.datastore.KvDataStore.KvRecord;
 import org.corfudb.runtime.exceptions.DataCorruptionException;
 import org.junit.Test;
